requirements.txt
asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
Pillow==2.6.1
django-currentuser==0.5.0
psycopg2-binary
awscli
gunicorn
django-rest-swagger==2.2.0
stripe==2.58